Bracket picking progression

+++ backend/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import UUID, JSON, Integer, String, ForeignKey, Text, Boolean
+from sqlalchemy import UUID, JSON, Integer, BigInteger, String, ForeignKey, Text, Boolean
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, deferred
 import uuid
 
@@ -12,13 +12,13 @@ class DBUser(Base):
     username: Mapped[str] = mapped_column(String(15), nullable=False)
     password: Mapped[str] = deferred(mapped_column(String(255), nullable=False))
     displayName: Mapped[str] = mapped_column(String(15), nullable=False)
-    createdEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
+    createdEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
 class DBRefreshToken(Base):
     __tablename__ = "refresh_token"
 
     token: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    expiryEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
+    expiryEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
     userId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("user.id"))
 
 class DBPool(Base):
@@ -28,9 +28,9 @@ class DBPool(Base):
     title: Mapped[str] = mapped_column(String(31), nullable=False)
     creatorId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("user.id"), nullable=True)
     bracketSourceId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("bracket_source.id"), nullable=False)
-    createdEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
-    startEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
-    endEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
+    createdEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    startEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    endEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
     active: Mapped[bool] = mapped_column(Boolean, nullable=False)
     settings: Mapped[JSON] = mapped_column(JSON)
 
@@ -43,7 +43,7 @@ class DBPick(Base):
     userId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("user.id"), nullable=False)
     poolId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("pool.id"), nullable=False)
     current: Mapped[bool] = mapped_column(Boolean, nullable=False)
-    pickEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
+    pickEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
 class DBBracketSource(Base):
     __tablename__ = "bracket_source"
@@ -52,13 +52,13 @@ class DBBracketSource(Base):
     name: Mapped[str] = mapped_column(String(31), nullable=False)
     fetchUrl: Mapped[str] = mapped_column(Text, nullable=False)
     lastFetch: Mapped[JSON] = mapped_column(JSON, nullable=True)
-    lastFetchEpoch: Mapped[int] = mapped_column(Integer, nullable=True)
+    lastFetchEpoch: Mapped[int] = mapped_column(BigInteger, nullable=True)
 
 class DBParticipant(Base):
     __tablename__ = "participant"
     
     userId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("user.id"), primary_key=True)
     poolId: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("pool.id"), primary_key=True)
-    joinedEpoch: Mapped[int] = mapped_column(Integer, nullable=False)
+    joinedEpoch: Mapped[int] = mapped_column(BigInteger, nullable=False)
     hidden: Mapped[bool] = mapped_column(Boolean, nullable=False)
     balance: Mapped[int] = mapped_column(Integer, nullable=False)
